+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/service/enforcetheorder/EnforcementOrderServiceTest.java
@@ -10,10 +10,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.pcs.ccd.domain.enforcetheorder.EnforcementOrder;
 import uk.gov.hmcts.reform.pcs.ccd.entity.ClaimEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.enforcetheorder.ConfirmEvictionEntity;
+import uk.gov.hmcts.reform.pcs.ccd.entity.confirmeviction.EvictionEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.enforcetheorder.EnforcementOrderEntity;
 import uk.gov.hmcts.reform.pcs.ccd.event.EventId;
-import uk.gov.hmcts.reform.pcs.ccd.repository.enforcetheorder.ConfirmEvictionRepository;
+import uk.gov.hmcts.reform.pcs.ccd.repository.enforcetheorder.EvictionRepository;
 import uk.gov.hmcts.reform.pcs.ccd.repository.enforcetheorder.EnforcementOrderRepository;
 import uk.gov.hmcts.reform.pcs.ccd.service.DraftCaseDataService;
 import uk.gov.hmcts.reform.pcs.ccd.service.PcsCaseService;
@@ -46,7 +46,7 @@ class EnforcementOrderServiceTest {
     @Mock
     private EnforcementTypeStrategyFactory strategyFactory;
     @Mock
-    private ConfirmEvictionRepository confirmEvictionRepository;
+    private EvictionRepository evictionRepository;
     @InjectMocks
     private EnforcementOrderService underTest;
 
@@ -184,7 +184,7 @@ class EnforcementOrderServiceTest {
         underTest.confirmEviction(CASE_REFERENCE);
 
         // Then
-        verify(confirmEvictionRepository).save(any(ConfirmEvictionEntity.class));
+        verify(evictionRepository).save(any(EvictionEntity.class));
     }
 
     @Test
